requirements.txt
paho-mqtt>=1.3.1
python-can>=1.5.2
